Put trait definition in common crate
