+++ common/statistics/Cargo.toml
@@ -9,7 +9,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = { version = "0.1.51" }
+log = "0.4"
 reqwest = "0.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
+sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "chrono"]}
 thiserror = "1"
+tokio = { version = "1.19.1", features = [ "time" ] }
