+++ common/statistics/src/lib.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 use error::StatsError;
 
 pub mod api;
+pub mod collector;
 pub mod error;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
