+++ service-providers/network-requester/src/statistics/mod.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-mod comm;
+mod collector;
 mod error;
 
-pub use comm::{ServiceStatisticsCollector, StatisticsSender, StatsData};
+pub use collector::ServiceStatisticsCollector;
